Changed all floats to doubles because FloatStream doesn't exist and the small difference in memory and time is negligible. Also add method for handling incoming signals to a node

+++ src/GraphNetwork/Global/GraphNetwork.java
@@ -62,7 +62,7 @@ public class GraphNetwork {
         return n;
     }
 
-    public Signal createSignal(final Node sendingNode, final Node recievingNode, final float strength)
+    public Signal createSignal(final Node sendingNode, final Node recievingNode, final double strength)
     {
         activeNextNodes.add(recievingNode); // every time a signal is created, the network is notified of the reciever
         return new Signal(sendingNode, recievingNode, strength);
